// codigo-fonte/frontend/src/api/models/create-business-case-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IgescConecta.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateBusinessCaseRequest
 */
export interface CreateBusinessCaseRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateBusinessCaseRequest
     */
    'name'?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof CreateBusinessCaseRequest
     */
    'originsBusinessCasesIds'?: Array<number>;
}