// codigo-fonte/frontend/src/api/models/list-business-case-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IgescConecta.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Filter } from './filter';

/**
 * 
 * @export
 * @interface ListBusinessCaseRequest
 */
export interface ListBusinessCaseRequest {
    /**
     * 
     * @type {number}
     * @memberof ListBusinessCaseRequest
     */
    'pageNumber'?: number;
    /**
     * 
     * @type {number}
     * @memberof ListBusinessCaseRequest
     */
    'pageSize'?: number;
    /**
     * 
     * @type {Array<Filter>}
     * @memberof ListBusinessCaseRequest
     */
    'filters'?: Array<Filter>;
}