// codigo-fonte/frontend/src/api/models/filter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * IgescConecta.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Op } from './op';

/**
 * 
 * @export
 * @interface Filter
 */
export interface Filter {
    /**
     * 
     * @type {string}
     * @memberof Filter
     */
    'propertyName'?: string;
    /**
     * 
     * @type {Op}
     * @memberof Filter
     */
    'operation'?: Op;
    /**
     * 
     * @type {any}
     * @memberof Filter
     */
    'value'?: any;
}